requirements.txt
accelerate
sentence-transformers
transformers
pypdf

langchain-text-splitters
langchain-community
langgraph
langchain-huggingface
langchain-core
langchain-ollama
langchain-chroma
psycopg[binary,pool]
langgraph-checkpoint-postgres

chromadb

tavily-python

flask
werkzeug
PyJWT